fix: handle too many notification windows opening
if windows get stacked up beyond the extent of the window fall back
to overlapping them from the bottom of the window. This is
unlikely/should not be happening but is the best solution for now.
Alternatively in future we can close all open windows if too many are
open