fix: don't analyse flutter sdk packages (#58)
Running analysis in the flutter sdk packages leads to 100s of diagnostic errors which slows down neovim significantly. This PR excludes this folder from analysis. Potentially in the future the analysis server can be active since this code should not really be generating that many errors in the first place.