fix(documentChanges): allow document renames
The lsp setting `renameFilesWithClasses` allows renaming a file if the
containing class is names the same as the file and the class name
changed, unfortunately nvim core does not explicitly enable
workspace.workspaceEdit.documentChanges which is required to trigger
this functionality so I've added it in here.

It should probably eventually be sent as a PR to core.