feat(syntax): include improved dart syntax file
vim's syntax file for dart is pretty simple and doesn't include much
syntax highlighting. The dart-vim-plugin on the other hand does a much
better job but comes with a whole bunch of unwanted logic in the land of
treesitter + lsp.

Why do we need syntax anyway if we have treesitter?
The floating windows for the LSP still use vim's syntax for syntax
highlighting not treesitter so to get rich syntax highlighting in the
lsp hover popups this plugin now ships with a copy of dart vim's syntax
file.

This will eventually be deleted once the hover doesn't need syntax
any more.