feat(log): add focus option for log window
Added a new configuration option `focus_on_open` to control whether the
log window should be focused when opened. Updated the default
`open_cmd` to `botright 30vnew` because this is default option.
Modified the `open_win` function to switch back to the previous window
if `focus_on_open` is set to false.