Fix "new Flutter Version" error on FlutterRun (#495)
* Improve type definitions for module executable

This commit is basically a small code-cleanup for the `executable.lua`
module, but it mainly adds Lua Doc comments to some public functions.

* Fix flutter version/welcome banner lua error

Executing any `flutter` command (including `flutter debug_adapter` and
`flutter debug-adapter`) may cause the `New Flutter Version` or
`Welcome to Flutter` banner to be printed to STDOUT. This is problematic,
because it interferes with `nvim-dap`s ability to parse RPC calls (obviously,
because the banners are not formatted to look like a valid RPC call).
`nvim-dap` is unable to properly detect the `Content-Length` header of
the first RPC call (because of the banner/s) -> `error(...)` gets called
-> user receives an error when trying to run their project.

This commit intends to fix that, by detecting and clearing any banners,
before the debug adapter is started.